requirements.txt
amadeus==10.1.0
python-dotenv==1.0.1
requests==2.32.3
twilio==9.2.3
python-dotenv==1.0.1
